After running Nightly  the time of Regimes-Speedup was 91.2 min while main was 91.5 min. While make-regime! that contains pareto-regimes takes 1012.34s in Regimes-Speedup, in main it takes 1073.72s. During profiling there is much more time spent in pareto-regimes since the exprs-to-branch-on is now being calculated during this function. This makes pareto-Regimes take 30.55s in Regimes-Speedup but only takes 12.74s in main. I'm unsure where exactly the speedup can be found in the profiling but it is faster I think.