@pavpanchekha 2024-07-09
I think the win is in `exprs-to-branch-on`: from 70s to 18s (4x faster!). But overall regimes time is longer—curious.