@pavpanchekha 2024-07-09
Ok, I think I know why. The number of calls went up, from 109 555 to 111 004. I think that's because over time there are fewer alts and regimes itself is called fewer times. One way to see this is to look at "Distance on a great circle". It gets faster, yes, but it also goes from 23 to 30 output alts. Or look at individual `regimes` calls; there's one that goes from 158 → 1 alt, and in `main` it involves 8 calls (8 exprs to branch on) but in this PR it makes 12 calls.

I could kind of be convinced to merge this as is anyway. The code is simpler and it does make the "Distance on a Great Circle" faster so maybe it's fine.