@pavpanchekha 2024-07-09
Another benefit of merging as is is that I have this idea to make regimes even faster and this is more compatible with that idea...